GPL License (As I understand it)
I can confirm that the versioninfo files in these mod folders show "Shadowing" or KeeperRL discord as the author and that "Shadowing" is also displayed on the mods page in game as author.

I can't prove the history of these files or the way the graphics have been put together but I spoke to Shadowing about an hour ago on discord as follows:

Shadowing — Today at 12:56 PM
ok pushed it. It contains beta for FromTheDepths

Soft Monster — Today at 1:14 PM
Thanks

Soft Monster — Today at 1:28 PM
I could mark you mods are licensed under GPL as free to use and modify but not free to claim as property

Shadowing — Today at 1:28 PM
that good

Soft Monster — Today at 1:32 PM
I can insert the GPL license file into your mod folders if that is helpful
This is what  I do
https://github.com/SoftMonster/KeeperRLCommunityResources/blob/master/Mods/Alpha36/A36BonusMod/COPYING.txt
______________

So here is the license as I understand it.