Add files via upload
Fem!Keepers have their staff hand two pixels to the right of Mas!Keepers.  This leads to their staff sticking to the back of their hand, which looks tacky.  As far as I could tell, the only solution was to either edit the vanilla tile offset (in vanilla files) or to shift the whole sprite; sprites thus shifted two pixels left.  This creates a noticeable offset on the starting menu and with their shadows, but is the best solution I know of.