Update workshops_menu.txt
As far as I can tell, if something's in the shop with certain stats, it has to be in all shops with those same stats.  KRL utterly refused to find anything (including the Rock it wanted to replace the unfindable BONUS_LongSpear and BONUS_Mace with) when the LongSpear had different shop stats than earlier.  Not sure if I fixed a typo by accident or if it was a mod-support issue, but wev.  Very much tested and works.