Fix tech mismatches
Turns out that "goblin_lore" was being called for but not in the tech definitions or allotment.  Since underscores aren't required in tech definitions, added "goblin lore" to the tech defs & allotment, and standardized 'em in the creatures, immigration, & spells.  Also removed a similarly undefined (but not crash-causing, since immigration wasn't calling for it) "gnomish lore" and gave gnomes "automatons", which it turns out was how one recruited gnomes anyway.