Fix BONUS_tech_mismatches
Same problem as with TechAbsorb, only also one-off lawful-immigrants who arrive with tech that the Church doesn't get.  We'll look into ways to get elephants & sprites in the future, but for now, I *hate* having things that crash the mod and trigger the stack trace.  Offending immigration commented out, Absorber keeper fixed as with TechAbsorb.